+++ pynenc/conf/config_pynenc.py
@@ -17,6 +17,8 @@ class ConfigPynenc(ConfigPynencBase):
         The state backend class to use.
     :cvar str serializer_cls:
         The serializer class to use.
+    :cvar str arg_cache_cls:
+        The argument cache class to use.
     :cvar str runner_cls:
         The runner class to use.
     :cvar bool dev_mode_force_sync_tasks:
@@ -33,6 +35,7 @@ class ConfigPynenc(ConfigPynencBase):
     broker_cls = ConfigField("MemBroker")
     state_backend_cls = ConfigField("MemStateBackend")
     serializer_cls = ConfigField("JsonSerializer")
+    arg_cache_cls = ConfigField("DisabledArgCache")
     runner_cls = ConfigField("DummyRunner")
     dev_mode_force_sync_tasks = ConfigField(False)
     max_pending_seconds = ConfigField(5.0)