+++ pynenc/serializer/__init__.py
@@ -1,5 +1,6 @@
 from pynenc.serializer.base_serializer import BaseSerializer
+from pynenc.serializer.constants import ReservedKeys
 from pynenc.serializer.json_serializer import JsonSerializer
 from pynenc.serializer.pickle_serializer import PickleSerializer
 
-__all__ = ["BaseSerializer", "JsonSerializer", "PickleSerializer"]
+__all__ = ["BaseSerializer", "JsonSerializer", "PickleSerializer", "ReservedKeys"]