+++ pynenc/serializer/json_serializer.py
@@ -1,21 +1,9 @@
 import builtins
 import json
-from enum import StrEnum
 from typing import Any
 
 from pynenc.serializer.base_serializer import BaseSerializer
-
-
-class ReservedKeys(StrEnum):
-    """
-    Defines keys reserved for internal use in the serialization process.
-
-    This enumeration ensures that specific keys are uniquely identifiable and not confused with user-defined data.
-
-    :cvar ERROR: Reserved key for storing error information in serialized data.
-    """
-
-    ERROR = "__pynenc__std_py_exc__"
+from pynenc.serializer.constants import ReservedKeys
 
 
 class DefaultJSONEncoder(json.JSONEncoder):